+++ mephisto/abstractions/architects/router/deploy/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "server",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js"
+  },
+  "keywords": [],
+  "author": "",
+  "engines": {
+    "node": ">= 7.6.0"
+  },
+  "dependencies": {
+    "async-lock": "1.0.0",
+    "body-parser": "^1.19.0",
+    "express": "^4.17.1",
+    "multer": "^1.4.2",
+    "request": "2.82.0",
+    "ws": "4.0.0"
+  }
+}
