+++ mephisto/abstractions/architects/router/deploy/static/index.html
@@ -1,0 +1,25 @@
+<!--
+
+Copyright 2017-present, Facebook, Inc.
+All rights reserved.
+
+This source code is licensed under the license found in the
+LICENSE file in the root directory of this source tree.
+
+-->
+
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Crowdsourcing Task</title>
+    <script src="wrap_crowd_source.js"></script>
+  </head>
+
+  <body>
+    <noscript>JS is required</noscript>
+    <div id="app" style="overflow: hidden"></div>
+    <script src="bundle.js"></script>
+  </body>
+</html>
