+++ mephisto/data_model/test/utils.py
@@ -24,7 +24,7 @@ import json
 
 from mephisto.abstractions.providers.mock.mock_provider import MockProviderArgs
 from mephisto.server.blueprints.mock.mock_blueprint import MockBlueprintArgs
-from mephisto.server.architects.mock_architect import MockArchitectArgs
+from mephisto.abstractions.architects.mock_architect import MockArchitectArgs
 from mephisto.data_model.task_config import TaskConfigArgs
 from mephisto.core.hydra_config import MephistoConfig
