+++ test/core/test_operator.py
@@ -17,7 +17,10 @@ from mephisto.data_model.test.utils import get_test_requester
 from mephisto.data_model.constants.assignment_state import AssignmentState
 from mephisto.core.local_database import LocalMephistoDB
 from mephisto.core.operator import Operator
-from mephisto.server.architects.mock_architect import MockArchitect, MockArchitectArgs
+from mephisto.abstractions.architects.mock_architect import (
+    MockArchitect,
+    MockArchitectArgs,
+)
 from mephisto.core.hydra_config import MephistoConfig
 from mephisto.abstractions.providers.mock.mock_provider import MockProviderArgs
 from mephisto.server.blueprints.mock.mock_blueprint import MockBlueprintArgs
